requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
boto3==1.20.49
botocore==1.23.49
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.4
colorama==0.4.4
ConfigArgParse==1.5.3
cryptography==36.0.1
cx-Oracle==8.3.0
Django==4.0
djangorestframework==3.13.1
docutils==0.18.1
Flask==2.0.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==1.5.3
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==0.10.0
Mako==1.2.0
MarkupSafe==2.1.1
msal==1.17.0
msgpack==1.0.3
netifaces==0.11.0
numpy==1.22.1
pandas==1.3.5
psutil==5.9.0
pycparser==2.21
pycryptodome==3.10.1
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
pywin32==303
PyYAML==6.0
pyzipcode==3.0.1
pyzmq==22.3.0
requests==2.27.1
roundrobin==0.0.2
s3transfer==0.5.1
SimpleParse==2.2.2
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
Werkzeug==2.0.3
xmltodict==0.12.0
zope.event==4.5.0
zope.interface==5.4.0
